+++ tests/MapTest.php
@@ -10,7 +10,7 @@ use Innmind\Immutable\{
     Set,
     Sequence,
 };
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class MapTest extends TestCase
 {