+++ tests/Set/DeferTest.php
@@ -12,7 +12,7 @@ use Innmind\Immutable\{
     Sequence,
     SideEffect,
 };
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class DeferTest extends TestCase
 {